Fix PSR-4 compliance for OsNotSupportedException
The class name had a typo causing the class to be called
OsNotSupportedExeception thus not matching the filename.

This could be considered a breaking change, for autoloaders that don't
ignore classes that don't conform to PSR-4. For those that do, it's not
really breaking, because currently the code would crash because it
can't load the class. *shrug*